Renamed the safe-insets package to android-runtime. Renamed the main group to androidlayouttranslator from xmltoxib.
