+++ android-runtime/src/main/AndroidManifest.xml
