+++ android-runtime/src/main/equivalents/swift/equivalents.swift.yaml
