+++ android-runtime/src/main/java/com/lightningkite/androidruntime/safeInsets.kt
@@ -1,4 +1,4 @@
-package com.lightningkite.safeinsets
+package com.lightningkite.androidruntime
 
 
 import android.os.Build
