+++ android-runtime/src/main/res/values/public.xml
