+++ test-project/build.gradle
@@ -42,7 +42,7 @@ android {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:1.5.30"
-    implementation project(':safe-insets')
+    implementation project(':android-runtime')
 
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
